issue-492 test into custom-parser.test.js file
